requirements.txt
Django
dj-database-url
django-configurations

django_installed_apps_generator
django_redis_pushpull_worker

psycopg2
psycopg2cffi
